Add Rust to REAMDE.md and to elapsed time comparison
